requirements.txt
kornia==0.6.0
opencv-python==4.2.0.34
matplotlib
pytorch-lightning==1.6.4
setuptools==59.5.0
test_tube
imageio==2.19.3
imageio-ffmpeg